Attend issue #5: fingerprints computation

+++ tidyscreen/moldyn/moldyn_utils.py
@@ -1,3 +1,5 @@
+import warnings
+warnings.filterwarnings("ignore")
 import shutil
 import subprocess
 import glob
@@ -6,6 +8,7 @@ from tidyscreen.GeneralFunctions import general_functions as general_functions
 import sys
 import pandas as pd
 from tidyscreen import tidyscreen
+
 
 def prepare_md_initial_files(assay_folder,complex_pdb_file,mol2_lig_parm,frcmod_lig_parm,solvent,min_steps,dynamics=1):
     ## Prepare the corresponding 'tleap' input file
@@ -294,13 +297,14 @@ def run_tleap_input(assay_folder,input_file):
     command = f'{tleap_path} -f {assay_folder}/{input_file}' 
     subprocess.run(command, shell=True, capture_output=True, text=True)
     
-def perform_minimization(assay_folder,ligand_prefix,solvent):
+def perform_minimization(assay_folder,ligand_prefix,solvent,inform=1):
     # Compute mol2 with Sybyl Atom Types - for compatibility with RDKit and Meeko
     pmemd_path = shutil.which('pmemd.cuda')
     # Run first minimization procedure
     try:
         command = f'{pmemd_path} -O -i {assay_folder}/min1.in -o {assay_folder}/min1.out -p {assay_folder}/complex.prmtop -c {assay_folder}/complex.inpcrd -r {assay_folder}/min1.crd -ref {assay_folder}/complex.inpcrd' 
-        print(f"Runnning 'min1' for ligand: '{ligand_prefix}'")
+        if inform == 1:
+            print(f"Runnning 'min1' for ligand: '{ligand_prefix}'")
         subprocess.run(command, shell=True, capture_output=True, text=True)
     except Exception as error:
         print(error)
@@ -308,7 +312,8 @@ def perform_minimization(assay_folder,ligand_prefix,solvent):
     if solvent == "explicit":
         # Run second minimization procedure
         command = f'{pmemd_path} -O -i {assay_folder}/min2.in -o {assay_folder}/min2.out -p {assay_folder}/complex.prmtop -c {assay_folder}/min1.crd -r {assay_folder}/min2.crd' 
-        print(f"Runnning 'min2' for ligand: '{ligand_prefix}'")
+        if inform == 1:
+            print(f"Runnning 'min2' for ligand: '{ligand_prefix}'")
         subprocess.run(command, shell=True, capture_output=True, text=True)
     
 def write_mmgbsa_input(assay_folder):
@@ -368,17 +373,32 @@ def compute_MMPBSA(assay_folder,traj_input):
     
     return assay_folder, f"{assay_folder}/mmgbsa_DECOMP.out"
     
-def clean_MMPBSA_files(assay_folder):
-    patterns_to_clean = ['*.inpcrd','*.prmtop','*.in','_MMPBSA*','*.crd','*frcmod','*.mol2','min*.out','*.log']
+def clean_MMPBSA_files(target_dir):
+    patterns_to_retain = ['renum.csv','_pose_']
     
-    for pattern in patterns_to_clean:
-        files_to_delete = glob.glob(os.path.join(assay_folder, pattern))
-        
-        for file_path in files_to_delete:
-            try:
+    # Construct a list of files to retain
+    files_to_retain = []
+    for pattern in patterns_to_retain:
+        for filename in os.listdir(target_dir):
+            if pattern in filename:
+                files_to_retain.append(filename)
+
+    # Delete file if not marked to be retained
+    for filename in os.listdir(target_dir):
+        if filename not in files_to_retain:
+            file_path = os.path.join(target_dir, filename)
+            if os.path.isfile(file_path):
                 os.remove(file_path)
-            except Exception as e:
-                pass
+    
+    
+    # for pattern in patterns_to_clean:
+    #     for filename in os.listdir(target_dir):
+    #         if pattern in filename:
+    #             for pattern_to_exclude in patterns_to_retain:
+    #                 if pattern_to_exclude not in filename
+    #                     file_path = os.path.join(target_dir, filename)
+    #                 if os.path.isfile(file_path):
+    #                 os.remove(file_path)
 
 def renumber_mmgbsa_output(assay_folder,decomp_file,receptor_filename):
     # Get a dictionary (resname_resnum) for the original receptor
@@ -394,10 +414,11 @@ def renumber_mmgbsa_output(assay_folder,decomp_file,receptor_filename):
 
     ## Process both dictionaries and output the final .csv file
     # This will out a .csv file with components and renumbered residues values
-    prepare_mmgbsa_renumbered_output(receptor_sequence_dict,decomp_file_dict,assay_folder)
+    decomp_csv_file_renum = prepare_mmgbsa_renumbered_output(receptor_sequence_dict,decomp_file_dict,assay_folder)
     # This will out a .csv file with components and amber residues values
-    prepare_mmgbsa_original_output(receptor_sequence_dict,decomp_file_dict,assay_folder)
+    decomp_csv_file = prepare_mmgbsa_original_output(receptor_sequence_dict,decomp_file_dict,assay_folder)
 
+    return decomp_csv_file,decomp_csv_file_renum
 
 def parse_mmgbsa_output(filename):
     with open(filename,'r') as input_file:
@@ -420,10 +441,10 @@ def parse_mmgbsa_output(filename):
     
     return file_resname_resnum_dict
 
-
 def prepare_mmgbsa_renumbered_output(dict1,dict2,assay_folder):
     # This will write a .csv file with the components with renumbered residue numbers
-    with open(f'{assay_folder}/mmgbsa_DECOMP_renum.csv','w') as output_file_renum:
+    renumbered_file = f'{assay_folder}/mmgbsa_DECOMP_renum.csv'
+    with open(renumbered_file,'w') as output_file_renum:
         counter = 1
         for key, value in dict2.items():
             resname,resnum = dict1[int(key)].split('_')
@@ -437,9 +458,12 @@ def prepare_mmgbsa_renumbered_output(dict1,dict2,assay_folder):
 
             output_file_renum.write(record)
 
+    return renumbered_file
+
 def prepare_mmgbsa_original_output(dict1,dict2,assay_folder):
     # This will write a .csv file with the components with renumbered residue numbers
-    with open(f'{assay_folder}/mmgbsa_DECOMP.csv','w') as output_file:
+    original_file = f'{assay_folder}/mmgbsa_DECOMP.csv'
+    with open(original_file,'w') as output_file:
         counter = 1
         for key, value in dict2.items():
             resname,resnum = dict1[int(key)].split('_')
@@ -452,6 +476,8 @@ def prepare_mmgbsa_original_output(dict1,dict2,assay_folder):
                 counter += 1
 
             output_file.write(record)
+
+    return original_file
 
 def create_md_assay_registry(db,docking_assay_id,docking_pose_id):
     # Try to create the 'dynamics_registries' table if it does not exist
@@ -474,5 +500,5 @@ def create_md_assay_registry(db,docking_assay_id,docking_pose_id):
     
     return last_assay_id
 
-def  create_md_assay_folder(assay_folder):
+def create_md_assay_folder(assay_folder):
     os.makedirs(f"{assay_folder}",exist_ok=True)
